refactor: changes jira-get-statuses to statuses + other platform support
