+++ src/commands/index.ts
@@ -1,8 +1,8 @@
 export const commands = new Set([
     'files',
     'help',
+    'issues',
     'jira-projects',
     'jira-statuses',
-    'jira-get-issues',
     'prepare',
   ]);
